Adding click events to the MainMenu Rocket GUI
Added a utility to streamline binding listeners for Rocket events,
making it much more JavaScript-like. The RocketListener template class
can be used to bind a member function of an object to an event.

Added click listeners for all the options in the Main Menu using the
RocketListener helper class. The only thing left is keyboard handling
and reselection sounds when changing options.

Added the base RCSS class from the libRocket site to grant default CSS
properties to RML elements.