Added RocketControls to build, added Menu keyboard support
Re-introduced keyboard support for the MainMenu state. Also fixed a bug
in which the Rocket<->SDL key mapping wasn't being initialized.

Added the RocketControls lib to the build. Currently this isn't being
used, but it most likely will be used for the Options menu in the
future. Added a CMake module specifically for finding the
RocketControls library+includes and adding them to the build.