Remaking the in-game menu using libRocket
This change removes the old Guichan-based menu and begins the addition of a new libRocket-based menu. Corresponding RML and RCSS files have been added, and the Home Menu has been changed to initialize a pane using a Rocket context.

The MenuShell class has been changed to hold the Rocket context for the menu state, as well as the menu frame (sidebar and background). This class will also eventually handle responses to list events. The MenuShell now only holds pointers to the RocketContext and ElementDocument, and these are reference counted by the Rocket library and deleted accordingly. The copy constructor and destructor of MenuShell increment and decrement references respectively. As a result, MenuShell instantiation (i.e. from MainMenu and TileEngine) no longer leaks memory.

GraphicsUtil has been changed to provide a factory method for Rocket contexts.

The EdenRocketBindings class was added to create and manage listeners created for a document. MainMenu::bindAction has been merged into this class since those binding helpers will be useful for every Rocket UI in the game.

A view model was added to expose the Player Data to the HomePane in a Rocket DataGrid. The result is that the character summary remains dynamically populated. An ImageFormatter was also added to wrap image paths in <img> tags.

This commit unfortunately marks a loss in functionality, as the menu prototype needs to be recreated from the ground up. However, implementation of a new prototype should be much easier, as the menu code is no longer bloated with widget creation and management.