Refactoring common menu code
Reintroducing the MenuState concept for the menu to hold common menu
state code (handling the rendering and action parsing for Rocket, for
instance).

The MenuShell now sends the currently active menu state events when a
sidebar option is clicked. The next commit will bring back other menus.

Also fixed a bug where the menu shell would move in front of the active
pane when clicked, since the background of the shell was not
transparent. To fix this, I split the menu shell into a background
document and a shell widget document.