Adding Data Menu, fixing some bugs
Added back the data menu along with its previous functionality. At this
point, the menu is roughly back to where it was before in terms of
functionality. The Guichan version of the menu also had mockups of the
Item and Equip menus, so those will need to be added back in as soon as
possible.

While doing this, also fixed a nasty segfault that resulted from the
EdenRocketBindings. Basically, when the EdenRocketBindings goes out of
scope, it deletes the listeners that it allocated. But since they're
still attached, there is a segfault when libRocket tries to detach the
listeners and run their OnDetach functions. To ensure that this won't
happen again, the RocketListener detaches itself on destruction,
ensuring that it will never be attached after it is destroyed.

Added a deactivate method for the ExecutionStack to call whenever a
state is removed from the top of the stack (either by being popped or
by another state being pushed on top of it). Currently, this is used by
the menu states to hide windows from states below the current active
state.

Added a check in ImageFormatter to ensure that null or empty strings
don't result in img tags being created.

Made MenuShell a pointer managed by MenuState… I was pretty happy about
the way I implemented this before, but having get copied across menu
panes meant that each copy had its own menu state reference, and so
each instance of the menu shell had a different "active state". This
meant that sidebar clicks would always go to the wrong state. Now, a
single MenuShell instance is shared by all menu states, and so there
are no longer any synchronization issues.

Added code to PlayerData to properly clear out old data before a new
save game is loaded in.