Converting DebugConsoleWindow to RML, removing EDWT
Removing the ED Widget Toolkit, created as an extension to the Guichan
widgets. All the extra stuff that was in EDWT is taken care of easily
by Rocket's event handling and by RCSS (such as text centering). This
reduces code bloat tremendously.

To ensure that functionality isn't lost, the DebugConsoleWindow was
rewritten to use an RML document like the rest of the GUI elements.
Some of the input in TileEngine was filtered to ensure that it won't
conflict with the debug console when that window is visible.

Fixing RocketSDLInputMapping to send text input commands in addition to
regular key press events, so that text input boxes would work correctly.