Fixing two GUI bugs.
Fixed an issue in which GUI isn't ready to render during transition:

On activation, a GameState's Rocket context is now updated to
ensure that the GUI is ready immediately. The result of this is
that transition states now present a finished GUI for transitions. This is most apparent when using a BlendState to transition to a menu state.

Fixing issue in which modifier keys don't properly work.

This issue was primarily due to lack of parentheses when using binary AND to test the keyboard modifier state. The result was that modifier keys (ctrl, alt, command, etc.) would never get sent to the Rocket GUI.