ScriptFactory touch-up, documentation corrections
Slight refactoring of one of ScriptFactory's private members to reduce
coupling and the number of parameters needed when loading an NPC script.

Ran Doxygen and found a series of @param tags that didn't match the
actual parameter lists for their functions and corrected them. Also
changed the NPCScript documentation to (hopefully) read more clearly.

Finally, altered the Doxyfile to output warnings for undocumented entities, and to pipe them to a warning file. Over time, I will be whittling away at the warnings until every class and method is clearly documented.