Adding a menu pane that allows viewing of inventory.
Also refactoring the menus to pull up the concept of a main
"paneDocument" to the MenuState superclass. With the MenuState taking
ownership of the menu pane document, the MenuState now exposes the
ability to turn the sidebar of the menu on/off depending on if there is
anything to show in the sidebar list or not.