Adding embedded Lua script support to Rocket menus
Created a script handler factory that the game context now registers
with Rocket when the game starts. This script handler factory will
create handlers for the various events that can be inlined into the RML
(onload, onclick, onfocus, etc).

The result of this is that we can embed Lua scripts into the GUI and
hopefully make the menus much more customizable through Lua in the
future.

To facilitate this work, the MainMenu and MenuState classes now create
Schedulers, just like the TileEngine does. The work that the TileEngine
did to calculate time between frames has been pulled up to the
GameState since it's now needed in any GameState that has a scheduler.

Also silencing a minor warning in Script.cpp.