Added support for using items!!
Woohoo! EDEn now supports the usage of items, which calls into script
functions organized in a very similar way to NPC script functions (i.e.
each file has a set of functions, and the script loads the file to
cache the functions and call them when an item is used).

Items have (up to) three functions for the contexts in which they are
used: onMenuUse, onFieldUse and onBattleUse for use in the Menu states,
tile engine and battle engine (future) respectively.

To facilitate this, moved ItemData into the GameContext instead of
making it a Singleton. Items are now only accessible through a game
context instance, which required piping the game context object all the
way through the PlayerData, down to the EquipSlots, which use the item
data.

At this point, I believe that there is only one singleton in the code,
GraphicsUtil, and since its initialization is explicitly fired on game
start, we can probably move it into the game context as well and
eliminate all singleton usage.

To test the item support out, I wired up the Item Menu to 'use' items
whenever they are clicked. Currently, clicking on a Potion 'uses' it,
causing a script message to print to console. In the future, item
scripts will be able to select a character to operate on if they need
to.

Finally, renamed savegamejson to savegame1 and changed the savegame and
item data to reserve item ID 0 to signify an invalid item.