Cleaning up some warnings
Reordering initializer lists to keep the field orders consistent with
class structure.

Also removed a glaring error where the Spritesheet frame list is
constructed with size 0 (it was being given NULL as a parameter because
it used to be a C-style array). This wouldn't have caused any visible
problems, but it was good to catch nonetheless!