Fixed regression in transition states
The term "timePassed" has a different meaning in TransitionState, where
it was originally used to represent ALL time passed since the start of
the transition. After changing this, I forgot to change the
transition's progress calculation to keep using the cumulative time
passed rather than the time between frames.

Though seeing the screen wobble and flash like that was kind of funky,
I'm glad to have transitions working the way that they should again. :)