Fixing bug when using item in debug mode.
The ItemScript's pointer to its parent Item was never initialized. This went by just fine without debug logs turned on, but when they were on we actually dereferenced this uninitialized pointer and thus a segfault could occur on item use.

Fixed ItemScript by initializing the Item pointer in the constructor, and then made the pointer a reference for safety, since it's always going to be explicitly initialized anyway.