Adding graceful failure for image load failure in GUI
Currently, if the RML or RCSS for a piece of the UI specify a bad image path, the Texture constructor throws and the program just dies. Fixed this by adding a "valid" flag for the texture to help indicate if it was successfully constructed. Our Rocket RenderInterface will do the right thing and clean up the texture, plus notify Rocket that the image load failed. The result is that bad image paths result in white quads instead of crashing the game. Far nicer, I think. :)