Adding item use support when in the field (TileEngine state)
Added GameStateType as an enum that allows the GameContext to determine
the current state type. This allows Item usage to differ based on the
currently active state.

Changed the shortcut bar and the item menu to call a generic "use"
function in items. Item::use will take the current GameState's
GameStateType into account when determining which script to call (menu
usage, field usage, or battle usage).

Also updated the Potion script to actually pop a speech dialog instead
of printing to log, since it just looks prettier that way.