Generalizing Item to Usable type
As the upcoming Skill type will rely on many of the same foundations as
Items (object prototypes with context-sensitive "use" scripts), a lot
of the Item and ItemScript code is reusable. Renamed Item to Usable to
account for this change. The expectation is that Skill will be a
subclass of Usable. As Item is functionally equivalent to Usable right
now, it has been turned into a typedef, thus leaving a lot of the code
that depends on Item relatively intact. Item.h was created as a
convenience header when Item itself is needed specifically.