Adding Skill Metadata and Skill Scripts
Updated the Metadata class to parse and store skill data. Also added
Skill scripts to the ScriptFactory and a test script.

Finally, updated item IDs to use a typedef called ItemId instead of
just an "int". Made the new type an unsigned int (since at this point 0
is considered invalid data, and the ID should never be negative).
Introduced SkillId for skills.