Adding Skills to Characters, adding skill menu
The skill menu was created to show a list of a character's skills. To facilitate this, the change also includes work in the HomeViewModel to handle character-dependent menus. When the Skills option is clicked in the sidebar, a character needs to be selected by clicking on one of the characters in the home pane. This work can later be extended for Equip and Status menus.

Fixed some issues with menu RCSS where the menu sidebar took up the width of the screen (though the visible area was correctly just the left 20% of the menu). This would cause the menu to appear unresponsive after a sidebar option was clicked, as the sidebar document was now in front of the menu pane. This change fixes the issue by moving some of the styles for the sidebar panel to the body of the menu shell document.

Also added some test skill data to the savegames. Assigned the two test skills, Heal and Fireball, to the character "kain". Added skill parsing and serialization into the Character class.