Adding Kain sprite and credit for Sithjester
Missed the addition of this spritesheet in a totally unrelated commit a
year ago, and just found this image in the repo with no credit attached.

Renamed the file appropriately and added proper credit to Sithjester.
Next commit will introduce slight character changes to actually make
use of this sprite for the main character.