Added new Player Character sprite
Removed a hack in which the player character loads and uses the npc1
spritesheet, since now the proper script/character-based mechanisms are
in place for setting the player sprite.

Actors are now "spriteless" on initialization, and so some guards were
added to prevent drawing or altering an uninitialized Sprite.

Resized the "kain" spritesheet by 1.5x to match the tile size and other
NPCs. Created an EDS file for the spritesheet. Also altered the "kain"
character archetype to use the new spritesheet.

Removed the map script's code to show the character; in the case of an
intro sequence this should be the responsibility of the chapter script.

Removed some unnecessary debug messages that were accidentally left in
with the last few commits.