Culling TODO items out of TODOLIST.h
Additional work for Issue #14. Removed several TODOLIST.h items and replaced them with issues opened on GitHub when they were still relevant.

Removed a lot of duplicate/obsolete items. One of the more interesting removals that doesn't have a matching issue is the reference-counting/memory pool functionality that I wanted ResourceLoader to have. While I can certainly see a lot of benefits to a resource management mechanism that routinely frees up resources that aren't being used, I don't see this as being a major priority until I can prove that memory is a major issue for EDEn. It may be that these items come back eventually, but for now I would like the backlog of issues to be focused on verified pain points for developers and/or players.

Removed some script function TODOs around the freezing/thawing of user input and NPC functioning. I feel that before blindly implementing these functions, a little bit more thought is required around how a cutscene in EDEn can be most easily scripted. As well, "freezing" an NPC or the PlayerCharacter requires determining exactly which functionality really needs to be "frozen", and what should still be happening. For example, the Player Character could be frozen by preventing all logic in PlayerCharacter::step from executing, but that would mean that a cutscene script can't manipulate the PlayerCharacter at all. It will be important to get a better feel for what should and shouldn't happen during a cutscene, before we can decide what functionality to expose to the scripting engine.