Adding more Music functionality to scripts
Added a new optional parameter to the stopMusic Lua function to supply
a fade time, in milliseconds. This allows the scripts to control how
the music stops, rather than just abruptly halting the music.

Added a "music man" NPC to the town to turn music on and off, testing a
variety of music functions. When spoken to while music is not playing,
the NPC turns on music. Otherwise, the NPC fades the music. To
facilitate the NPC's logic, a global isMusicPlaying function was also
added to the scripting engine.

Finally, I'm able to use Oren's music in our game. He made the town
song for the game years ago, and it's great to finally see it in use. :D