Adding build support for Visual Studio 2012
Generating a VS2012 project with CMake uncovered a few problems that had
to be fixed before compiling and running EDEn successfully using a
Visual Studio solution.

CMakeLists.txt was updated to link to the appropriate libraries
depending on whether the build is Release or Debug.
RocketScriptHandlerFactory was updated to match the latest libRocket
EventListenerInstancer interface.
PlayerData::shortcutList initializes its shortcutList vector with empty
shortcuts rather than 0 (which doesn't make sense since it's a vector of
Shortcuts).

In addition to the work to get EDEn building in VS2012, I also updated a
couple of places to use the Rocket::Core::Element::SetAttributes method
for setting multiple attributes at once.