Adding settings file to remove hardcoded settings.
EDEn will now read a settings.ini file to get the game settings and user options. All settings will go through this file from now on, instead of being either hardcoded or #defined in the code. A future check-in will expose an Options menu that allows the user to manipulate the settings.