Game creates a default settings.ini on startup.
From now on, the settings.ini file is automatically created when it does not exist. This means that developers can specify their own preferred settings without having to change source code (and risk checking in their own preferences!). The settings.ini file is no longer checked into source and is now ignored, as it is auto-generated by the game.