Adding sound/music options to Options Menu
Added two checkboxes to the options menu for music and sound, and
hooked them up to the current Settings.

Made Settings into a non-static class so that multiple instances of
Settings could be created (for caching old Settings to allow for
Revert/Cancel behaviour in the options menu in the future).

Currently, the checkboxes don't render their icons correctly and are
thus represented by simple red/green boxes. A future fix will add
proper checkbox rendering in the options menu.