Adding ability to revert changes in settings menu, initial resolution change support
Settings now holds a snapshot, which is meant to represent the way the
Settings looked at the point of loading/saving (i.e. how it looks "on
disk"). This allows us to revert changes made in the settings menus by
the user if they aren't happy with the changes made.

Also started making some separations around video settings so that
resolution/fullscreen mode changes are properly reflected with a window
change. Unfortunately, SDL 1.2 destroys the OpenGL context associated
with the window whenever the screen is disposed, so all the textures go
away. As a result, this change doesn't actually allow users to alter
the resolution or window mode on the fly. This work will have to be
completed after the move to SDL2.