Game now correctly responds to resolution changes
The TileEngine and the Rocket contexts driving the game's GUI now
appropriately resize in the event of a resolution change. The
TileEngine changes size on activation (after popping a Game State that
potentially altered settings), and all Rocket Contexts are now tracked
in a registry, allowing the Graphics library to resize all active
contexts on any resolution change.

Until the move to SDL2, resolution change will still cause textures
(and everything else in the OpenGL context) to be destroyed. However,
once this move is completed and resolution change is enabled, these
changes will make updates to video settings more seamless.