[#51142391] Adding basic camera slide support
This change adds the CameraSlider, which is a Coroutine that
asynchronously moves the TileEngine's camera from one point to another.
in a straight line.

The change also introduces the ability for the TileEngine to lock the
camera to any arbitrary Actor, instead of just the PlayerCharacter. The
TileEngine can also release this lock when the Camera is being
manipulated (e.g. by the CameraSlider).