Adding Lua hooks for manipulating the camera
[#51142391]
Exposing three new script functions from the TileEngine, allowing a
script writer to slide the camera, make it follow an Actor, or release
its existing lock on an Actor.
The Slide functionality requires a destination, but can optionally take
an origin, which defaults to the current camera focal point if not given
by the writer. Speed is also an optional parameter.
Exposed new functionality from the TileEngine to manipulate the camera
and to get its current point of focus. Also altered ScriptUtilities to
support non-integer numeric parameters (floats/doubles).