Fix camera slide delay between start and finish
[Finishes #51142391]

Optimized the CameraSlider to clip parts of the slide that would not
result in actual camera movement (clamped areas near map edges). As a
result, all the frames of the slide where the camera wouldn't move are
now optimized out. As a result, the slide starts right away and doesn't
continue when it's finished moving the camera.

Also added slides to the Chapter 1 Script for testing. This uncovered a
design issue in which the Camera slide was not synchronous to the
script, which allowed multiple slides to execute simultaneously. This
would lead to undefined behavior. From now on, the slide is always
synchronous and the script always waits on it. The Chapter 1 Script now
slides the camera to the right, then back to the player sprite, and then
locks back to the player for verification of the slider functionality.