Merging options menu changes into master
[Finishes #50939287]

Merging options menu changes into the main branch. Also fixing an issue
caused by the merge between the camera slide work and the options menu
work. Specifically, recalculation of the map offset now needs to check
initially if a map is defined, since it is called on activation and the
TileEngine may not have a map up every time that it is activated.

As well, when a player enters a new area, the camera now snaps to the
player immediately by default.