Migrating EDEn code to SDL 2.0
[#49268621]

Migrates the codebase to SDL 2.0 by making use of new APIs where needed.
Specifically:
- The PlayerCharacter keyboard state check now uses scancodes instead of
keycodes as required
- The GraphicsUtil startup creates an OpenGL context and a window to
render to. The context lives for the lifetime of the app and may survive
past the lifetime of the window it was initially created for (if the
window is destroyed for a video mode change during the execution of the
game)
- Added some missing cleanup for SDL_Mixer resources

Also altered CMakeLists.txt to reference SDL 2.0 for Windows. The
changes needed for non-Windows machines to pick up SDL 2.0 will be
checked in with the next commit.