Preceding all private members with "m_"
Some code review feedback on Reddit indicated to me that there are a
lot of places in the codebase where the names of member variables are
the same as parameter names. A frequent example is in class
constructors: a parameter is assigned to a member with the same name
and then used in the constructor function's body. I'm now preceding all
private class members with "m_" to add some clarity.