Fixing Quest script function issues
The conversion to the new parameter format for EDEn Lua functions
caused some of the Quest script functions to break. In particular,
usages of "id" as a parameter name needed to be replaced with "name"
when adding quests. Querying for quests using a subquest path (e.g.
"quest/subquest/subsubquest") wasn't properly converted, so new
overloads of "isStarted" and "isComplete" were created to support
subquest paths.