Adding Dialogue completion cursor
[#49250757]

Adding a dialogue completion cursor that appears when dialogue can be
advanced. The cursor is controlled using the RCSS class
"dialogueComplete", which is applied to the entire dialogue box
document on completion of the current line of dialogue. Custom styling
can be applied based on this class. A "readyNote" class was added to
the RCSS to take advantage of this.