Adding script callbacks for Map Triggers
[#49379841]
Lua scripts can now bind to the current map and listen for map trigger
events being fired by the EntityGrid. Callbacks are stored using Lua's
registry table, and the registry index is encapsulated in a
MapTriggerCallback object for ease of passing the callback to the
TileEngine.

The scripts have been updated with an example of this new feature. The
next commit will break the two existing maps into a larger number of
maps so that not all the features are running at once (which has been
historically difficult to debug).