Pushing more archetype data into save data
The objective for character archetypes, in general, should be a single
usage for character creation. After that, the character data should be
all that is loaded when an ongoing game is loaded. Otherwise, it is
impossible to alter character data that exists in the archetype, such
as the portrait or spritesheet data.

Future work will push Aspect data into the save game, but only when
Aspect data itself is removed from the archetype and only referenced by
an Aspect ID.