Decoupling ExecutionStack from GameContext
After putting some significant work into the Skill system that EDEn
will be using, I realized I was passing GameContext around to nearly
every new class I was writing because it was the only way to access
certain objects. That's when I realized the usage of the Context object
was getting
out of hand.

Many game states will need access to common instances of services, like
the ScriptEngine or the ExecutionStack. And making these accessible via
dependency injection is a great way to control how these instances are
initialized, used and tested. However, bundling these objects in a
Context object, while innocuous at first, makes it too easy to start
passing the entire world to every object. And while we aren't at that
point yet with EDEn, the GameContext is in a lot of places and it's
important to keep evaluating its usage and making sure that it doesn't
become a "God object".

This commit removes the ExecutionStack instance from the
GameContext and ensures that it gets set explicitly on each GameState
as it is processed by the one ExecutionStack instance in the game.
Since GameState's access to ExecutionStack is protected, this means
only GameState objects can actually access the ExecutionStack. That
means that each of the MenuState subclasses does the work of navigating
between the states, instead of deferring to a view model, which makes
the MenuState objects act more like Controllers (as they should be).
Places that originally
took GameContext objects but really only needed the ExecutionStack now
get the ExecutionStack directly. This has led to some more granular
parameter passing, but now several classes have way less access to
things they don't need, which is how it should be.