Adding Skill usage to Character data
Characters now unlock skills based on skill usage. This means that
instead of the Character simply having a list of skills previously
unlocked, there is a map of skills to successful usage counts. If these
counts are high enough compared to the mastery requirement of the
skill, the skill is considered available for use.