Optimizing skill tree validation a bit
Aspect::validateSkillTree made use of an additional map in order to
help the validation process along. The extra map not only needlessly
takes more memory, but also impacts the readability of the code. I've
made some changes to remove extra data structures, and to better
clarify some of the useful structures using typedefs.