Changing RocketListeners to pass Events by ref
As a small addition to the change to RocketListener in 0549ff8 to use
lambdas, RocketListener now always passes events by ref instead of by
pointer. The pointer passing was a workaround to allow calls to the
handlers even with no event (by passing nullptr). Instead, now handlers
that don’t care about the event just don’t take the event parameter.
With the lambda notation, supporting parameterless handlers is cleaner.